Member 기능 1차 완료
SNS측은 이제 추가하도록 하겠음
아마 기존 Member 테이블에 타입을 추가하여
google인지 facebook인지 email인지 확인할 수 있도록 하겠음

+++ JSP&Spring/dokky/src/main/java/com/kitri/dokky/member/MemberController.java
@@ -146,23 +146,32 @@ public class MemberController {
 	@ResponseBody
 	@RequestMapping("/deleteMember") 
 	public String deleteMember(@RequestParam("mem_id")String mem_id,
+			@RequestParam("mem_pw")String mem_pw,
 			HttpSession session) {
-
-		service.deleteMember(mem_id);
-
-		Member deleteMemberData = service.getMember(mem_id);
+		
+		Member getMemberData = service.getMember(mem_id);
 
 		String result = "";
-
-		if(deleteMemberData != null) {
-			//result = "error/error";
-			result = "{'result':'fail'}";
+		
+		if(getMemberData.getMem_pw().equals(mem_pw)) {
+			service.deleteMember(mem_id);
+			
+			Member deleteMemberData = service.getMember(mem_id);
+			
+			
+			if(deleteMemberData != null) {
+				//result = "error/error";
+				result = "{'result':'fail'}";
+			}
+			else {
+				sessionInvalidate(session);
+				
+				//result = "redirect:/main/home";
+				result = "{'result':'success'}";
+			}
 		}
 		else {
-			sessionInvalidate(session);
-
-			//result = "redirect:/main/home";
-			result = "{'result':'success'}";
+			result = "{'result':'passNo'}";
 		}
 
 		return result;
